Move SchemaCommandResponse to submodule
